requirements.txt
streamlit
pandas
matplotlib
spacy
pdfminer.six
keybert
sentence-transformers
